move MonthHeader to inside PlantCalendarList
