--- components/MonthHeader/styles.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const MonthsContainer = styled.div`
-  display: grid;
-  grid-template-columns: repeat(12, 1fr);
-  /* gap: 0.75rem; */
-  width: 100%;
-  justify-items: center;
-`;
